Add files via upload
Unfortunately, could only complete part 1. Not used Binary Trees before, so this was a steep learning curve. The code wasn't properly adding sizes I believe past a certain point due to duplicate directory names. It allowed for part 1 to be completed but not part 2. If I get more time I'll return to this to try and complete.